requirements.txt
streamlit
pandas
openpyxl
matplotlib
python-dotenv
langchain
langchain-groq
